pyproject.toml
[project]
name = "rpa-project-template"
version = "0.1.0"
description = ""
authors = [
    {name = "XBF2016",email = "272216402@qq.com"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.13"
dependencies = [
]

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
playwright = "^1.44.0"
pandas = "^2.2.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.7.1"
black = "^24.4.2"
ruff = "^0.4.9"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
